coap_io.c,net.c: Fix head of node queue re-transmit timeouts
The node queue uses time deltas between each entry in the queue for
retransmissions.  The head entry does not have a time delta.

When the head of the head of the queue is taken off for retransmission
the next in the queue needs to have the time delta updated to be an
absolute value.  However, this does not happen if the current head has had
the time offset fore set to 0 and so the retransmit timer fires more frequently
than expected.

There are 2 ways to fix this - not maintain time deltas between queue
entries, or not set the head entries time to 0.

Doing the latter means that code that relies on the deltas (I can only
find the test suite) continue to work before calling coap_retransmit().

See Issue #197 for further information.