src/resource.c: Non Confirmable observe requests not gettng out if CON pending
If an non-confirmable observe response was triggered, and there was already a
CON message outstanding, then the observe response would not get sent in error.

[Example being a observe response following a Ping 0.00 that has not yet had
a reply]

The fix is to check whether the observe request is non-confirmable and there
are less than COAP_OBS_MAX_NON already sent to allow it to get sent.