configure.ac: add option for gcov flags
Gcov[1] is a tool to investigate test code coverage. To collect
runtime statistics, certain compiler options must be set, i.e.
-fprofile-arcs, -ftest-coverage, and --coverage. Currently, only
the Gnu C Compiler is known to support the gcov instrumentation.

This change adds the configure option --enable-gcov that checks if
the compiler supports the aforementioned compiler flags and enables
these.

[1] https://gcc.gnu.org/onlinedocs/gcc/Gcov.html