Makefile.am: make headers co-installable to previous version(s)
The recently, in the last developemnt circle done TLS related changes to
libcoap had make the library incompatible to the old API and thus to the
old used header files.

The (new) library is already installed with the API suffix and by this
is not conflicting with previous libcoap versions.
Doing the same by this and following commits for the needed header of
libcoap.