configure.ac: fix evaluate logic for library suffix
If a user isn't explicitly disabling the DTLS support for libcoap or
requesting the usage of tinydtls then instead the configure script needs
to detect at least the development files for GnuTLS or OpenSSL.
If not the script must break and give up with an error message.

Because of some wrong logic within the evaluation if one of both TLS
libraries was found configure was *not* stopping with an error message.

The logic to check the possibilities of an automatic usage of the TLS
libraries was different, the check for GnuTLS was correct but on
OpenSSL a wrong check condition chain was used, it should and must have
been the same on both libraries. This commit is fixing this.
As all conditions must the true in theses check to set the correct
variables which are used later within configure.

But the sanity check at the end was also not completely correct and
failing so configure was not running into the desired error break.