GnuTLS If a bad key was used, it is not possible to get a new session
src/coap_gnutls.c:

The GnuTLS server session information held in endpoint->hello was getting
polluted when a bad PSK was being used, and so a new session attempt with a
good PSK would fail. The only way to get around this was to restart the
coap-server.

OpenSSL handles this by initiating the handshake response (DTLSv1_listen()),
creating a new coap session and then using the new session to continue with
the SSL handshake (SSL_accept()) until the SSL session is set up.

For GnuTLS, there is no initial handshake initiation as everything is done
inside gnutls_handshake().  To get around this, in the post_client_hello
function a flag is set in g_env->seen_client_hello so that coap_dtls_hello()
knows that that a SSL session is being set up, but not yet complete.
This flag is cleared when coap_dtls_hello() returns 1 indicating that the
interim client_hello has been seen and it is safe to duplicate the session.

The do_gnutls_handshake() function has been cleaned up.

The coap_dtls_send() function has been updated to send any outstanding data
if the SSL session has just completed.