GnuTLS Server: Handle session setup failures following a session setup failure
src/coap_gnutls.s:

If a DTLS server session setup fails (invalid / missing certificate etc.)
and the client has sent some more of the DTLS setup exchange packets before
receiving the TLS alert packet closing things down, any subsequent new DTLS
sessions will fail.

Caused by the g_env environment getting partially polluted with information
with the subsequent DTLS setup exchange packets.

The primary fix is for coap_dtls_hello() is to clear out g_env if there is a
ClientHello packet failure (error return from do_gnutls_handshake()) as well
as if the packet is not of the form of a ClientHello, then it is simply
ignored.

(This may be what is being seen in #300)