net.c: Start new coap_context with a random message identifier
This is done by calling prng_init() with a non-zero value in coap_startup().

Change 7fb4ace2b709994bbbebab4c2d367ac76140eee9 modified the parameter to
prng_init() from a computed value to 0 when the code was re-organized.

This change uses the current time in microseconds as the seed to prng_init().

include/coap2/coap_time.h:

Add in coap_ticks_to_rt_us() support for lwIP and Contiki

src/net.c:

Update call to prng_init() in coap_startup() to use time in microsecs as the
seed parameter

See #373 where the issue was raised.