coap_debug.c: Fix format truncation warning
When COAP_DEBUG_BUF_SIZE is smaller than the buffer size for holding
temporary strings in coap_show_pdu(), the compiler will emit a
warning as snprintf() truncates the output string. Moreover, this
will result in a missing terminating zero.

This fix adjusts the temporary buffer size if COAP_DEBUG_BUF_SIZE
is very small (i.e., on Contiki), and sets the last character in
outbuf to \000 when necessary.