net.c: Fix memory access after session is freed
Make sure that session is not freed off in one of the callbacks while
testing the socket flags in the epoll logic.

Wrap the code with a coap_session_reference(session) and
coap_session_release(session) and remove all the interim wrappers.

Issue found when coap_read_session() internally does a coap_session_release().