coap_gnutls.c: Fix potential spin in DTLS retransmit timeout loop
coap_dtls_get_timeout() may continue to return 'now' if the underlying
GnuTLS library gets confused by duplicate packets and so clears out
the data is to be retransmitted flag.  If there is a DTLS timeout retry, the
underlying logic does not reset 'last retransmit' on retrying the re-send
of the (now) non-existent data, so causing a spin in the coap_write() logic.

src/coap_gnutls.c:

This fix is to return 'now+dtls_timout' on the second and subsequents attempts
to return 'now'.

include/coap2/coap_dtls.h:

Make use of the new #defines that #353 is using so that this code compiles.