coap_debug.c: Make the PDU debug output amount fully configurable by a -D option
There are 2 output buffers - outbuf[] and buf[].  buf[] is limited in size
but used for printing out the data portion of the PDU before getting included
in outbuf[].

Make the printing of the data portion of the PDU go straight into outbuf[] to
get around the buf[] limitation.

Now the maximium amount of the printing out of the PDU can be controlled by
CPPFLAGS="-DCOAP_DEBUG_BUF_SIZE=xxxx" as a parameter to ./configure, where
xxxx is the output buffer size (e.g. use 4096).