DTLS: Sessions unnecessarily being re-handshaked
Found with GnuTLS, believed not to be an issue with OpenSSL and likely
(but not reproduced) to be an issue with MbedTLS.

In coap_io_prepare_io() the DTLS layer is checked to see if there should be
a (handshake) packet re-transmitted.  The logic checks to see if there is
a timeout and if so causes it to be re-transmitted.

Unfortunately this check was also being done when not in the DTLS handshake
phase and GnuTLS was returning 'timed out' based on the time since the last
handshake transmission.  As there was nothing being sent or received, when
doing the retransmission, this caused a re-handshake.

The fix is to only check for timeouts / re-transmits when in the handshake
phase.  This change does rigorous checking of state.

src/coap_io.c:

Only check timeouts / re-transmits when in Handshake phase.
[There is also a cpu savings as most sessions will not be in the
Handshake phase.]

src/coap_gnutls.c:
src/coap_mbedtls.c:
src/coap_openssl.c:

coap_dtls_get_timeout() assert if not handshaking
coap_dtls_handle_timeout() assert if not handshaking

Note: It is much easier to observe time when checking for this issue
if #441 changes are included.