CMakeLists.txt: Put BUILD_SHARED_LIBS option before add_library()
To build shared libraries, either the BUILD_SHARED_LIBS=ON option can be
defined globally for building, otherwise defining BUILD_SHARED_LIBS as an
OPTION in CMakeLists.txt works providing the OPTION is defined before the
add_library() call.

This fix is to move the OPTION BUILD_SHARED_LIBS to before the
add_library() call.

NOTE: Shared Library support is not currently available for Windows as the
appropriate functions are not currently exported.

BUILDING:

Document the Windows limitation and tidy up a bit.