net.c: Relax TCP CSM Block-Wise-Transfer tests
If the initial coap_send() PDU initiates a TCP/TLS connection, the PDU is
delayed until the CSM packets are exchanged.  If the initial PDU contains
a BLOCK1 or BLOCK2 option, the PDU gets dropped as Block-Wise-Transfer in the
CSM has not been seen yet.  Seen if '-b X' option is used for coap-client.

The check is now only when the session has fully established and the
situation only logged and the PDU Not dropped.