Allow network interface to be specified when joining multicast group
This change renames `coap_join_mcast_group()` to `coap_join_mcast_group_netif()`
and adds an extra `unsigned int intf_index` parameter to specify the network
interface, which is passed to `setsockopt()`. If it is zero, then it falls back
to resolving `::` as before.

A `coap_join_mcast_group` macro is defined to maintain backwards API
compatibility.