net.c: Stop reliable sessions returning ACK 0.00
The ACK 0.00 is being returned by the server for reliable conections and the
client is then reporting that CON 0.00 is invalid (CON, NON, ACK and RST are
all resported as CON in a received reliable PDU).

coap-client coap+tcp://127.0.0.1/async?1 -v9

demonstrates the issue.

src/net.c:

Update no_response() to return RESPONSE_DROP for both reliable and
non-confirmable if the application handler returns with the response code set
to 0.00 in the PDU. This then generally simplifies the code around when
calling no_response().

Check for valid multicast responses.

If the response is confirmable and piggybacked and being dropped, convert the
response to an empty ACK so the client does not continue to retry sending the
request.