net.c: Fix side effect from merging 2 PRs
Merging #552 and #554 into develop caused Observe CON 2.05 initial block
response not to get an ACK from the client and so the Observe response kept
getting re-transmitted and caused other CON traffic to get queued up.

Troubleshooting highlighted an issue when a PUT/POST was updating the contents
of a resource and an Observe was active sending many payloads in a body.

src/net.c:

Make sure that an ACK is sent (if appropriate) when working with a block based
response.

examples/coap-server.c:

Give dynamic data associated with a resource a reference count, and only free
off the data when the reference count has dropped to 0.