mcast block2: Support mcast request with block2 responses
The first block response is unicast, and the client requests the further blocks
using unicast.  https://tools.ietf.org/html/rfc7959#section-2.8

However, the hash for the server lookup of incoming data is based on
coap_addr_tuple_t which includes the local address - which changes between
receipt of mcast and unicast requests.  This causes 2 sessions to be set up,
and the subsequent request/responses are treated as individual block random
access causing mismatch of information by the client.

Fix is to hash only on the remote address (IP + port) and local port for
servers.