libraries: Rename libcoap-x to libcoap-x-notls and make libcoap-x default linker
Make the old installed libcoap-2.{so,la,a,pc} libcoap-2-notls.{so,la,a,pc}, and
make the installed libcoap-2.{so,la,a,pc} be a set of soft links to installed
libcoap-2.{tlstype}.{so,la,a,pc}.

Thus installed libcoap-2.{so,la,a,pc} will be based on the latest (D)TLS
library build.

Update the linking documentation.