coap_dtls_pki_sni_callback_t: Include coap_session_t in the parameter list
coap_dtls_psk_sni_callback_t already includes the session, but unfortunately
coap_dtls_pki_sni_callback_t has no way of signalling back to the application
(e.g. using coap_session_set_app_data()) as to which PKI information was used
for the (D)TLS session set up.

Including the session does break the API, but very few people are likely to
be using the validate_sni_call_back to modify the PKI information for the
(D)TLS session and so it is a small coding change if needed.